Refactor global entry picking flows
- replace GlobalFieldPicker call sites with a shared chip-plus-dialog GlobalEntryField
- add browse/create tabs to the new GlobalEntryDialog, including simple name-only creation and glaze combination composition
- remove the old picker components and migrate coverage to dialog-based tests
- update Bazel sources so focused web tests and eslint track the new picker structure

Validation:
- rtk bazel test //...
- rtk bazel build --config=lint //...
- source env.sh && gz_build
